[#26] refactor: Add service layer to card controller
card 관련 컨트롤러에 서비스 계층을 추가하였습니다.

+++ BE/todo/src/main/java/team9/todo/repository/CardRepository.java
@@ -6,5 +6,5 @@ import team9.todo.domain.Card;
 import java.util.List;
 
 public interface CardRepository extends CrudRepository<Card, Long> {
-    public List<Card> findAllByUserAndColumnType(int user, String columnType);
+    public List<Card> findAllByUserAndColumnType(long user, String columnType);
 }
